fix(ui): 修复 deno check 拉入 Nuxt composable 导致的类型检查失败
