+++ noj-ui/composables/useSearch.ts
@@ -1,35 +1,4 @@
-export type SearchEntityType =
-  | 'problem'
-  | 'user'
-  | 'community_post'
-  | 'community_comment'
-  | 'contest'
-  | 'submission'
-  | 'message'
-  | 'announcement';
-
-export type SearchType = 'all' | SearchEntityType;
-
-export interface SearchItem {
-  entity_type: SearchEntityType;
-  entity_id: string;
-  title: string;
-  highlight: string;
-  rank: number;
-  metadata: Record<string, unknown>;
-}
-
-export interface SearchState {
-  open: boolean;
-  query: string;
-  type: SearchType;
-  groups: Record<string, { items: SearchItem[]; has_more: boolean }>;
-  flatItems: SearchItem[];
-  hasMore: boolean;
-  page: number;
-  loading: boolean;
-  error: string | null;
-}
+export type { SearchEntityType, SearchItem, SearchState, SearchType } from '../utils/search-types.ts';
 
 export function useSearch() {
   const { api } = useApi();
